Added "Login with SAML bearer token" request
This request allows you to login with a SAML bearer token. Unfortunately,
Postman does not provide a way out of the box to provide a function to
gzip a string. The SAML bearer token needs to be gzipped and base64
encoded before it can be used. Due to not having gzip available, this
request requires you to gzip and base64 encode the token before providing
it.

Signed-off-by: J.R. Garcia <jrg@vmware.com>